Agent/DesktopPet/requirements_coding_agent.txt
# Additional requirements for Coding Agent features
# Install with: pip install -r requirements_coding_agent.txt

# AI Integration
google-generativeai>=0.3.0
python-dotenv>=0.21.0

# Git Integration (for Phase 2)
GitPython>=3.1.0

# Clipboard Monitoring (for Phase 3)
pyperclip>=1.8.0

# Code Analysis (for Phase 5)
pycodestyle>=2.10.0
radon>=5.1.0
